+++ src/test/java/com/example/demo/service/StudentServiceTest.java
@@ -21,7 +21,7 @@ public class StudentServiceTest {
 
     @Mock
     private StudentRepository studentRepository;
-    private StudentService studentService;
+    private StudentServiceImpl studentService;
     // Given
     private Student student = new Student(
         "test2",
@@ -32,7 +32,7 @@ public class StudentServiceTest {
     @BeforeEach
     void setUp() {
         MockitoAnnotations.openMocks(this);
-        studentService = new StudentService(studentRepository);
+        studentService = new StudentServiceImpl(studentRepository);
         // When
         studentService.addNewStudent(student);
     }
